Have Packfile3 follow the rule of 5
Trying to make memory management a bit saner. I don't like that this requires having 3 different constructors and 2 operators but I think it's an improvement over having a Cleanup function that you can easily forget to call. 

It might be possible to use the rule of 0 for this class and others if we make a class that can handle resource management for us, or maybe just by using a smart pointer. Need to read into the requirements of the rule of 0 a bit more first. This article is a good start: https://turingtester.wordpress.com/2015/06/27/cs-rule-of-zero/