Save progress
Some meshes (so far, building meshes) seem to interleave their render blocks by which material the renderblock is. The current way of reading render blocks where you assume all renderblocks of a submesh are adjacent to each other doesn't work with that. This was an attempt to fix that by determining which material each submesh was using and finding their renderblocks based on that. Doesn't work yet, may be the wrong approach. Saving progress for later attempts.