Save progress
This is an attempt to do chunk file reading (cchk/buildings). Using Nanoforge to visualize output. The code here is a hot mess and just a quick hack to use the StaticMesh class for chunk reading so I didn't have to make the nanoforge mesh viewing code more flexible for a simple test. 

So far have recreated previous results. Can read and view all of the submeshes but have two problems:
- Meshes are stacked on top of each other. Need to locate and read additional position/orientation data from file to put them in the correct positions.
- Meshes aren't using the right materials. This is more of a nanoforge problem but worth mentioning. Nanoforge doesn't fully support RFGs material system yet and so all the submeshes have the same textures. Will need to fix to make the buildings authentic to the game.