Fix bug in StaticMesh::ReadSubmeshData
Before this fix it would only return a subset of the vertex buffer, with the fix it returns the entire vertex buffer regardless of which submesh you requested. This is since all indices are referring to the entire vertex buffer rather than just some subset of it that a submesh might require. 

The main effect of this is that submeshes with an index > 0 look a bit better. This wasn't noticed before since the current supported mesh formats only use those submeshes as low lod meshes. So I thought they were just really bad looking low lod meshes.

I noticed the bug when I was testing out reading building meshes and saw that only the first submesh looked correct.