Add remaining RFG file formats to support table
Most of them are not supported but that doesn't mean nothing is known about them. Some like terrain zones (cterrain_pc) have had a great deal of them reverse engineered but a general purpose reader hasn't been added for them yet. At the time of writing Nanoforge draws terrain meshes in it's territory viewer but it uses a shortcut to locate the mesh data and extract it.