Alternative packfile unpacking method
Alternative method that is faster at the cost of higher memory usage. In many cases the increase in memory usage is negligible. In the worst case it'll use as much memory as the largest vpp_pc (~6.7GB for RFGR). Disabled by default so it still works works with computers with small amounts of RAM.