Remove packfile alphabetical sorting
This caused a mismatch in the order of primitives in str2_pc and asm_pc files. Any code that repacked asm_pc files in Nanoforge would have to correct for this making it more complicated. Instead opting to remove the sorting here and for anything that still requires sorting to do it with it's own copy of the file list. The Nanoforge file explorer already does this.