Fix bug in Packfile3::ExtractSingleFile()
Fixed case sensitivity bug. The target file wouldn't be found if there was a case mismatch in the provided name and name stored in the packfile.  

Also simplified the function a bit since I realized it was doing duplicate work of finding the entry when ::Contains() already provided an index.