Add and utilize defer macro for easier resource management
The defer statement is an idea taken from other languages like zig and go. It can be emulated in C++ using macros, with slightly different behavior than go defer statements. The defer macro lets you specify code to be run once the current scope ends. This makes temporary resource management code much nicer and clearer since you can put initialization and cleanup code write next to each other instead of on the far ends of a function. For example, lets say we allocated a buffer in a function and wanted to clean it up when the function returns. Instead of doing so at the end of the function, or possibly at each return statement, we'd do this:
```cpp
u8* buffer = new u8[4096];
defer(delete[] buffer);
```
Once the function goes out of scope the contents of the defer macro will be called. This is very convenient for code like that.