New experimental zone file reader
New experimental zone file reader that reads them in a similar way that the game does. This design uses the formats strengths. It loads the file into memory all at once then patches pointers to the different data sections. No parsing or conversion required. 

In theory this will result in simpler code and faster loading. Conversion to other formats would be the responsibility of tools using this library. This change is part of the changes being made to Nanoforge to use an editor data format internally and to keep RFG specific format details in importer/exporter functions.

The design for this and the Nanoforge editor format are still a work in progress and being evaluated for usefulness. The API for reading zone objects and properties in this new design could be improved.