Partial support for RFG steam edition static meshes
So far the differences are fairly minor. They added extra padding to some values in RFGR which get patched into relative pointers at runtime. I assume this is because they either were going to convert the PC version to 64bit, or had to convert it to 64bit for it to run on newer consoles. Still needs a lot more testing. Noticed there was a parsing error related to materials for the propane tank csmesh, so start there.